+++ src/App.tsx
@@ -1,21 +1,9 @@
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import { createGlobalStyle } from "styled-components";
 import { AuthProvider } from "./AuthContext";
 import reset from "styled-reset";
-import MapPage from "./pages/mapPage/MapPage";
-import Login from "./pages/Login/Login";
-import Signup from "./pages/Signup/Signup";
-import Layout from "./component/layout_navigation";
-import Portfolio from "./routes/portfolio";
-import Coupon from "./routes/coupon";
-import InvestList from "./routes/invest_list";
-import Return from "./routes/return";
-import CouponSelect from "./routes/coupon_select";
 
-import Home from "./routes/home";
-import ProjectDetailPage from "./pages/projectDetailPage/ProjectDetailPage";
-import FundingSuccessPage from "./component/fundingSuccess/FundingSuccess";
-
+import { RouterProvider } from "react-router-dom";
+import router from "./routes/router"; // router 그대로 가져옴
 const GlobalStyles = createGlobalStyle`
   ${reset}
 
@@ -83,40 +71,7 @@ const GlobalStyles = createGlobalStyle`
     }
   }
 `;
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Layout />,
-    children: [
-      { path: "", element: <Home /> },
-      { path: "portfolio", element: <Portfolio /> },
-      {
-        path: "/invest-list",
-        element: <InvestList />,
-      },
-      { path: "/map", element: <MapPage /> },
-      { path: "/map/:id", element: <ProjectDetailPage /> },
-    ],
-  },
-  {
-    path: "/coupon",
-    element: <Coupon />,
-  },
-  { path: "/return/:invest_id", element: <Return /> },
-  { path: "/coupon_select", element: <CouponSelect /> },
-  {
-    path: "/login",
-    element: <Login />,
-  },
-  {
-    path: "/signup",
-    element: <Signup />,
-  },
-  {
-    path: "/fundingSuccess",
-    element: <FundingSuccessPage />,
-  },
-]);
+
 function App() {
   return (
     <>
